requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
PyMySQL==0.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
requests-aws4auth==1.0
six==1.15.0
urllib3==1.25.10
